Add ci scripts for windows (#251)
* [skip-tests] Add ci scripts for windows

* [skip-tests] Prepare testing with sccache

sccache currently does not support the `-v` flag on MSVC. Also we cannot use it to test for supported compile flags as it will just break.

Finally on windows we want to compile with `--use-local-env` to avoid nvcc setting up the environment every time

* [skip-tests] Simplify the test workflow and add a windows run

* [skip-tests] For now replace sccache binary with custom one to get it to work with MSVC

* Fix `_LIBCUDACXX_UNREACHABLE` for MSVC and old nvcc

* Make libcu++ pass with MSVC

* Make thrust pass with MSVC

* [skip-tests] Make cub pass with MSVC

* Expand test matrix to include windows runs

Do not run the examples on windows as we do not have GPU runners with windows images yet

+++ cub/test/catch2_test_device_scan_iterators.cu
@@ -301,8 +301,8 @@ CUB_TEST("Device scan works complex accumulator types", "[scan][device]")
 
   custom_accumulator_t init{};
 
-  thrust::device_vector<custom_input_t> d_input(num_items, custom_input_t{1});
-  thrust::device_vector<custom_output_t> d_output{num_items, custom_output_t{nullptr, 0}};
+  thrust::device_vector<custom_input_t> d_input(static_cast<size_t>(num_items), custom_input_t{1});
+  thrust::device_vector<custom_output_t> d_output{static_cast<size_t>(num_items), custom_output_t{nullptr, 0}};
   thrust::device_vector<int> d_ok_count(1);
 
   auto index_it = thrust::make_counting_iterator(0);
